feat(2.2): Steps 1-3 — HirType i32/f64→Number + IntLiteral f64 + codegen + as
Step 1: HIR type unification
- HirType::I32/F64 merged into HirType::Number (hir.rs:252)
- from_ast_type: NumberType→Number (was I32)
- as_rust_type/Display: Number→number
- typeck: removed i32+f64→error, all arithmetic accepts Number
- IntLiteral(f64) upgrade: lexer/AST/HIR/TIR all i32→f64
- moveck is_copy_type: Number → Copy

Step 2: codegen f64 mapping
- Removed TYPE_I32, renamed TYPE_F64→TYPE_NUMBER
- hir_type_to_rust: Number→f64
- infer_tmp_type/tir_value_type: → TYPE_NUMBER
- Literal emission: {v}_f64 suffix

Step 3: as convergence
- check_as_cast: Number→Number rejected (identity cast)
- Same-type no-op retained for non-Number types
- All i32/f64 test assertions updated
- Integration tests: mix_numbers→OK, as_cast→rejected

Tests: cargo test --workspace 0 failures

+++ crates/trust_parser/src/lexer.rs
@@ -110,7 +110,7 @@ pub enum TokenKind {
     Null,    // §2.7
     Panic,   // §5.2 — Phase 4
     // 字面量
-    IntLiteral(i32),
+    IntLiteral(f64), // v2.0: f64（设计 §2.2）
     FloatLiteral(f64),
     StrLiteral(String),
     TemplateHead(String),
@@ -567,7 +567,7 @@ impl Lexer {
         }
         // v2.0: BigIntLiteral removed — 'n' suffix no longer supported
         let s: String = self.source[start..self.pos].iter().collect();
-        self.emit(TokenKind::IntLiteral(s.parse().unwrap_or(0)))
+        self.emit(TokenKind::IntLiteral(s.parse().unwrap_or(0.0))) // v2.0: f64
     }
 
     fn lex_ident(&mut self) -> TokenKind {
@@ -611,11 +611,11 @@ mod tests {
         assert!(ts.iter().any(|t| matches!(t, TokenKind::Let)));
         assert!(ts.iter().any(|t| matches!(t, TokenKind::Ident(s) if s == "x")));
         assert!(ts.iter().any(|t| matches!(t, TokenKind::NumberType)));
-        assert!(ts.iter().any(|t| matches!(t, TokenKind::IntLiteral(42))));
+        assert!(ts.iter().any(|t| matches!(t, TokenKind::IntLiteral(42.0))));
     }
     #[test]
     fn lex_number_literal_returns_int_token() {
-        assert_eq!(tokenize("42")[0], TokenKind::IntLiteral(42));
+        assert_eq!(tokenize("42")[0], TokenKind::IntLiteral(42.0));
     }
     #[test]
     fn lex_float_literal_returns_float_token() {
